commit 2

+++ src/main/java/com/eea/timetablesystem/Repository/UserRepository.java
@@ -4,9 +4,9 @@ import com.eea.timetablesystem.Model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+
 @Repository
-public interface UserRepository extends JpaRepository<User,Integer>
+public interface UserRepository extends JpaRepository<User, Long>
 {
     User findByEmail(String email);
-
 }
